chore: AgentProvider OO + shrink __all__ + tach merges
Wave 4.B for terok-executor: combined "finish W3.B OO" + API-surface
trim + tach module merges in a single PR.

**OO refactor (W3.B leftover).**  ``apply_provider_config(provider,
config, overrides)`` and ``build_headless_command(provider, *,
timeout, model, max_turns)`` were the textbook anaemic-domain pattern
— first arg ``provider: AgentProvider`` plus headless-only logic that
the provider's own descriptor fields drove.  Move both onto
``AgentProvider`` as ``.apply_config()`` and
``.build_headless_command()``, with the Claude / generic dispatch
helpers converted to private methods.  Move ``ProviderConfig`` and
``CLIOverrides`` into ``providers.py`` so the dataclass and its
methods live together.  ``resolve_provider_value`` (formerly in a
separate ``provider/config.py``) moves into ``providers.py`` too —
it's the internal helper ``AgentProvider.apply_config`` calls.

Free functions intentionally kept (they're cross-input or pure
helpers, not anaemic): ``assemble_container_env(spec, roster)``,
``ensure_vault_routes(cfg)``, ``seed_workspace_from_clone_cache(...)``.

**Tach merges.**  ``provider/config.py`` (49 lines, one function) and
``provider/headless.py`` (249 lines, two free functions + two private
helpers) fold into ``provider/providers.py`` — the descriptor and
its behaviour now share one module.  Tach module count 33 → 31.

**API surface trim.**  ``__all__`` shrinks 92 → 66 entries.  Removed
implementation-detail symbols no consumer (``terok``, tests, ACP
host-proxy) actually reads from the package root:

- ACP errors / internal types: ``ACPRoster``, ``AgentBindError``,
  ``AgentRosterCache``, ``ProbeError``
- credential schema internals: ``AuthProvider`` (one method shy of
  the public surface; reach via ``credentials.auth``),
  ``store_api_key``, ``extract_credential``,
  ``PHANTOM_CREDENTIALS_MARKER``, ``ConfigPatchError``
- doctor types re-export from sandbox (``CheckVerdict``,
  ``DoctorCheck``)
- container build internals: ``ImageSet``, ``INSTALLED_ENV_PATH``,
  ``l1_sidecar_image_tag``, ``render_l1_sidecar``,
  ``inject_agent_config``
- container env internals: ``ContainerEnvResult`` (used only as the
  return type of ``assemble_container_env``; reach via
  ``container.env`` for the type itself)
- roster internals: ``VaultRoute``, ``SidecarSpec``
- storage internals: ``get_task_storage``
- legacy CLI helpers: ``mounts_dir``, ``writable_config_path``
- the now-method versions: ``apply_provider_config``,
  ``build_headless_command``, ``collect_opencode_provider_env``
- removed orphan: ``CommandDef`` (terok-util owns it)

Docstring autorefs that pointed at the demoted symbols at the
``terok_executor.VaultRoute.X`` path are repointed at
``terok_executor.roster.VaultRoute.X``.  ``properdocs build --strict``
clean.

Pinned to terok-sandbox PR #354 (the W4.A counterpart) for the
duration of this PR — reverts to a released wheel on chain merge.
